misc refactorings
