+++ vine/src/distiller.rs
@@ -1,9 +1,6 @@
-mod pattern_matching;
-
 use std::mem::{replace, take};
 
 use ivm::ext::{ExtFn, ExtFnKind};
-use pattern_matching::Row;
 use vine_util::{
   idx::{Counter, IdxVec},
   unwrap_idx_vec,
@@ -13,7 +10,7 @@ use crate::{
   analyzer::usage::Usage,
   ast::{
     BinaryOp, Block, Builtin, ComparisonOp, DynFnId, Expr, ExprKind, GenericPath, LabelId, Local,
-    LogicalOp, Pat, PatKind, Span, StmtKind, Ty,
+    Pat, PatKind, StmtKind,
   },
   resolver::{Def, DefId, Resolver, ValueDefKind},
   vir::{
@@ -21,6 +18,9 @@ use crate::{
     VIR,
   },
 };
+
+mod control_flow;
+mod pattern_matching;
 
 #[derive(Debug)]
 pub struct Distiller<'core, 'r> {
@@ -100,64 +100,6 @@ impl<'core, 'r> Distiller<'core, 'r> {
     }
   }
 
-  fn new_stage(&mut self, layer: &mut Layer, interface: InterfaceId) -> Stage {
-    let id = self.stages.push(None);
-    layer.stages.push(id);
-    Stage {
-      id,
-      layer: layer.id,
-      interface,
-      header: Vec::new(),
-      declarations: Vec::new(),
-      steps: Vec::new(),
-      transfer: None,
-      wires: Counter::default(),
-    }
-  }
-
-  fn new_layer(&mut self) -> Layer {
-    let id = self.layers.push(None);
-    Layer { id, parent: None, stages: Vec::new() }
-  }
-
-  fn child_layer(&mut self, parent_stage: &mut Stage) -> (Layer, Stage) {
-    let mut layer = self.new_layer();
-    layer.parent = Some(parent_stage.layer);
-    let stage = self.new_unconditional_stage(&mut layer);
-    parent_stage.steps.push(Step::Transfer(Transfer::unconditional(stage.interface)));
-    (layer, stage)
-  }
-
-  fn root_layer(&mut self) -> (Layer, Stage) {
-    let mut layer = self.new_layer();
-    let stage = self.new_unconditional_stage(&mut layer);
-    (layer, stage)
-  }
-
-  fn finish_stage(&mut self, stage: Stage) {
-    let id = stage.id;
-    self.stages[id] = Some(stage);
-  }
-
-  fn finish_layer(&mut self, layer: Layer) {
-    let id = layer.id;
-    self.layers[id] = Some(layer);
-  }
-
-  fn new_unconditional_stage(&mut self, layer: &mut Layer) -> Stage {
-    let interface = self.interfaces.push(None);
-    let stage = self.new_stage(layer, interface);
-    self.interfaces[interface] =
-      Some(Interface::new(interface, layer.id, InterfaceKind::Unconditional(stage.id)));
-    stage
-  }
-
-  fn new_local(&mut self, stage: &mut Stage) -> Local {
-    let local = self.locals.next();
-    stage.declarations.push(local);
-    local
-  }
-
   fn distill_block(&mut self, stage: &mut Stage, block: &Block<'core>) -> Port {
     let mut result = Port::Erase;
     for stmt in &block.stmts {
@@ -198,6 +140,17 @@ impl<'core, 'r> Distiller<'core, 'r> {
       ExprKind![sugar || error || !value] => unreachable!("{expr:?}"),
 
       ExprKind![cond] => self.distill_cond_bool(stage, expr, false),
+      ExprKind::Do(label, block) => self.distill_do(stage, label.as_id(), block),
+      ExprKind::If(cond, then_branch, else_branch) => {
+        self.distill_if(stage, cond, then_branch, else_branch)
+      }
+      ExprKind::While(label, cond, block) => self.distill_while(stage, label.as_id(), cond, block),
+      ExprKind::Loop(label, block) => self.distill_loop(stage, label.as_id(), block),
+      ExprKind::Return(value) => self.distill_return(stage, value),
+      ExprKind::Break(label, value) => self.distill_break(stage, label.as_id(), value),
+      ExprKind::Continue(label) => self.distill_continue(stage, label.as_id()),
+      ExprKind::Fn(params, _, body) => self.distill_fn(stage, params, body),
+      ExprKind::Match(value, arms) => self.distill_match(stage, value, arms),
 
       ExprKind::Paren(inner) => self.distill_expr_value(stage, inner),
       ExprKind::Char(c) => Port::N32(*c as u32),
@@ -320,151 +273,6 @@ impl<'core, 'r> Distiller<'core, 'r> {
           lhs = next_lhs;
         }
         last_result
-      }
-
-      ExprKind::Do(label, block) => {
-        let local = self.new_local(stage);
-        let (layer, mut body_stage) = self.child_layer(stage);
-
-        *self.labels.get_or_extend(label.as_id()) =
-          Some(Label { layer: layer.id, continue_transfer: None, break_value: Some(local) });
-
-        let result = self.distill_block(&mut body_stage, block);
-        body_stage.set_local_to(local, result);
-
-        self.finish_stage(body_stage);
-        self.finish_layer(layer);
-
-        stage.take_local(local)
-      }
-
-      ExprKind::If(cond, then_branch, else_branch) => {
-        let local = self.new_local(stage);
-        let (mut layer, mut cond_stage) = self.child_layer(stage);
-        let (mut then_stage, mut else_stage) = self.distill_cond(&mut layer, &mut cond_stage, cond);
-
-        let result = self.distill_block(&mut then_stage, then_branch);
-        then_stage.set_local_to(local, result);
-
-        let result = self.distill_expr_value(&mut else_stage, else_branch);
-        else_stage.set_local_to(local, result);
-
-        self.finish_stage(cond_stage);
-        self.finish_stage(then_stage);
-        self.finish_stage(else_stage);
-        self.finish_layer(layer);
-
-        stage.take_local(local)
-      }
-
-      ExprKind::While(label, cond, block) => {
-        let (mut layer, mut cond_stage) = self.child_layer(stage);
-
-        *self.labels.get_or_extend(label.as_id()) = Some(Label {
-          layer: layer.id,
-          continue_transfer: Some(cond_stage.interface),
-          break_value: None,
-        });
-
-        let (mut then_stage, else_stage) = self.distill_cond(&mut layer, &mut cond_stage, cond);
-
-        let result = self.distill_block(&mut then_stage, block);
-        then_stage.erase(result);
-        then_stage.transfer = Some(Transfer::unconditional(cond_stage.interface));
-
-        self.finish_stage(cond_stage);
-        self.finish_stage(then_stage);
-        self.finish_stage(else_stage);
-        self.finish_layer(layer);
-
-        Port::Erase
-      }
-
-      ExprKind::Loop(label, block) => {
-        let local = self.new_local(stage);
-        let (layer, mut body_stage) = self.child_layer(stage);
-
-        *self.labels.get_or_extend(label.as_id()) = Some(Label {
-          layer: layer.id,
-          continue_transfer: Some(body_stage.interface),
-          break_value: Some(local),
-        });
-
-        let result = self.distill_block(&mut body_stage, block);
-        body_stage.erase(result);
-        body_stage.transfer = Some(Transfer::unconditional(body_stage.interface));
-
-        self.finish_stage(body_stage);
-        self.finish_layer(layer);
-
-        stage.take_local(local)
-      }
-
-      ExprKind::Return(value) => {
-        let value =
-          value.as_ref().map(|v| self.distill_expr_value(stage, v)).unwrap_or(Port::Erase);
-        let return_ = self.returns.last().unwrap();
-
-        stage.set_local_to(return_.local, value);
-        stage.steps.push(Step::Diverge(return_.layer, None));
-
-        Port::Erase
-      }
-
-      ExprKind::Break(label, value) => {
-        let value =
-          value.as_ref().map(|v| self.distill_expr_value(stage, v)).unwrap_or(Port::Erase);
-
-        let label = self.labels[label.as_id()].as_ref().unwrap();
-        if let Some(local) = label.break_value {
-          stage.set_local_to(local, value);
-        } else {
-          stage.erase(value);
-        }
-
-        stage.steps.push(Step::Diverge(label.layer, None));
-
-        Port::Erase
-      }
-
-      ExprKind::Continue(label) => {
-        let label = self.labels[label.as_id()].as_ref().unwrap();
-        let transfer = Transfer::unconditional(label.continue_transfer.unwrap());
-        stage.steps.push(Step::Diverge(label.layer, Some(transfer)));
-        Port::Erase
-      }
-
-      ExprKind::Fn(params, _, body) => {
-        let fn_local = self.new_local(stage);
-        let (layer, mut body_stage) = self.child_layer(stage);
-
-        self._distill_fn(&mut body_stage, fn_local, params, &**body, Self::distill_expr_value);
-
-        self.finish_stage(body_stage);
-        self.finish_layer(layer);
-
-        stage.take_local(fn_local)
-      }
-
-      ExprKind::Match(value, arms) => {
-        let local = self.new_local(stage);
-        let (mut layer, mut init_stage) = self.child_layer(stage);
-        let value = self.distill_expr_value(&mut init_stage, value);
-        let rows = arms
-          .iter()
-          .map(|(pat, expr)| {
-            let mut stage = self.new_unconditional_stage(&mut layer);
-            let result = self.distill_expr_value(&mut stage, expr);
-            stage.set_local_to(local, result);
-            let interface = stage.interface;
-            self.finish_stage(stage);
-            Row::new(pat, interface)
-          })
-          .collect();
-        self.distill_match(&mut layer, &mut init_stage, value, rows);
-        self.finish_stage(init_stage);
-        self.finish_layer(layer);
-        stage.take_local(local)
       }
     }
   }
@@ -618,6 +426,64 @@ impl<'core, 'r> Distiller<'core, 'r> {
     }
   }
 
+  fn new_stage(&mut self, layer: &mut Layer, interface: InterfaceId) -> Stage {
+    let id = self.stages.push(None);
+    layer.stages.push(id);
+    Stage {
+      id,
+      layer: layer.id,
+      interface,
+      header: Vec::new(),
+      declarations: Vec::new(),
+      steps: Vec::new(),
+      transfer: None,
+      wires: Counter::default(),
+    }
+  }
+
+  fn new_layer(&mut self) -> Layer {
+    let id = self.layers.push(None);
+    Layer { id, parent: None, stages: Vec::new() }
+  }
+
+  fn child_layer(&mut self, parent_stage: &mut Stage) -> (Layer, Stage) {
+    let mut layer = self.new_layer();
+    layer.parent = Some(parent_stage.layer);
+    let stage = self.new_unconditional_stage(&mut layer);
+    parent_stage.steps.push(Step::Transfer(Transfer::unconditional(stage.interface)));
+    (layer, stage)
+  }
+
+  fn root_layer(&mut self) -> (Layer, Stage) {
+    let mut layer = self.new_layer();
+    let stage = self.new_unconditional_stage(&mut layer);
+    (layer, stage)
+  }
+
+  fn finish_stage(&mut self, stage: Stage) {
+    let id = stage.id;
+    self.stages[id] = Some(stage);
+  }
+
+  fn finish_layer(&mut self, layer: Layer) {
+    let id = layer.id;
+    self.layers[id] = Some(layer);
+  }
+
+  fn new_unconditional_stage(&mut self, layer: &mut Layer) -> Stage {
+    let interface = self.interfaces.push(None);
+    let stage = self.new_stage(layer, interface);
+    self.interfaces[interface] =
+      Some(Interface::new(interface, layer.id, InterfaceKind::Unconditional(stage.id)));
+    stage
+  }
+
+  fn new_local(&mut self, stage: &mut Stage) -> Local {
+    let local = self.locals.next();
+    stage.declarations.push(local);
+    local
+  }
+
   fn distill_bin_op(&mut self, stage: &mut Stage, op: BinaryOp, lhs: Port, rhs: Port, out: Port) {
     let ext_fn = match op {
       BinaryOp::Concat => {
@@ -636,144 +502,6 @@ impl<'core, 'r> Distiller<'core, 'r> {
       BinaryOp::Rem => ExtFnKind::rem,
     };
     stage.steps.push(Step::ExtFn(ext_fn.into(), lhs, rhs, out));
-  }
-
-  fn distill_cond_bool(&mut self, stage: &mut Stage, cond: &Expr<'core>, negate: bool) -> Port {
-    if let ExprKind::Paren(inner) = &cond.kind {
-      return self.distill_cond_bool(stage, inner, negate);
-    }
-    match &cond.kind {
-      ExprKind![!cond] => {
-        let value = self.distill_expr_value(stage, cond);
-        if negate {
-          let wire = stage.new_wire();
-          stage.steps.push(Step::ExtFn(ExtFnKind::eq.into(), value, Port::N32(0), wire.0));
-          wire.1
-        } else {
-          value
-        }
-      }
-      ExprKind::Bool(b) => Port::N32((*b ^ negate) as u32),
-      ExprKind::Not(inner) => self.distill_cond_bool(stage, inner, !negate),
-      _ => {
-        let local = self.new_local(stage);
-        let (mut sub_layer, mut sub_stage) = self.child_layer(stage);
-        let (mut true_stage, mut false_stage) =
-          self.distill_cond(&mut sub_layer, &mut sub_stage, cond);
-        true_stage.set_local_to(local, Port::N32(if negate { 0 } else { 1 }));
-        false_stage.set_local_to(local, Port::N32(if negate { 1 } else { 0 }));
-        self.finish_stage(true_stage);
-        self.finish_stage(false_stage);
-        self.finish_stage(sub_stage);
-        self.finish_layer(sub_layer);
-        stage.take_local(local)
-      }
-    }
-  }
-
-  fn distill_cond(
-    &mut self,
-    layer: &mut Layer,
-    stage: &mut Stage,
-    cond: &Expr<'core>,
-  ) -> (Stage, Stage) {
-    if let ExprKind::Paren(inner) = &cond.kind {
-      return self.distill_cond(layer, stage, inner);
-    }
-    if let ExprKind::ComparisonOp(lhs, cmps) = &cond.kind {
-      if let [(op, rhs)] = &**cmps {
-        if matches!(rhs.kind, ExprKind::N32(0)) {
-          if *op == ComparisonOp::Ne {
-            return self.distill_cond(layer, stage, lhs);
-          } else if *op == ComparisonOp::Eq {
-            return swap(self.distill_cond(layer, stage, lhs));
-          }
-        }
-      }
-    }
-    match &cond.kind {
-      ExprKind![!cond] => {
-        let bool = self.distill_expr_value(stage, cond);
-        let interface = self.interfaces.push(None);
-        let true_stage = self.new_stage(layer, interface);
-        let false_stage = self.new_stage(layer, interface);
-        self.interfaces[interface] = Some(Interface::new(
-          interface,
-          layer.id,
-          InterfaceKind::Branch([false_stage.id, true_stage.id]),
-        ));
-        stage.transfer = Some(Transfer { interface, data: Some(bool) });
-        (true_stage, false_stage)
-      }
-      ExprKind::Bool(bool) => {
-        let true_stage = self.new_unconditional_stage(layer);
-        let false_stage = self.new_unconditional_stage(layer);
-        stage.transfer = Some(Transfer::unconditional(if *bool {
-          true_stage.interface
-        } else {
-          false_stage.interface
-        }));
-        (true_stage, false_stage)
-      }
-      ExprKind::Not(inner) => swap(self.distill_cond(layer, stage, inner)),
-      ExprKind::Is(value, pat) => {
-        let value = self.distill_expr_value(stage, value);
-        let true_stage = self.new_unconditional_stage(layer);
-        let false_stage = self.new_unconditional_stage(layer);
-        self.distill_match(
-          layer,
-          stage,
-          value,
-          vec![
-            Row::new(pat, true_stage.interface),
-            Row::new(&Pat { span: Span::NONE, kind: PatKind::Hole }, false_stage.interface),
-          ],
-        );
-        (true_stage, false_stage)
-      }
-      ExprKind::LogicalOp(op, a, b) => {
-        let (invert_out, invert_a, invert_b) = match op {
-          LogicalOp::And => (false, false, false),
-          LogicalOp::Or => (true, true, true), // A || B == !(!A && !B)
-          LogicalOp::Implies => (true, false, true), // A => B == !(A && !B)
-        };
-        let (mut a_true, mut a_false) = swap_if(invert_a, self.distill_cond(layer, stage, a));
-        let (b_true, mut b_false) = swap_if(invert_b, self.distill_cond(layer, &mut a_true, b));
-        let true_stage = b_true;
-        let false_stage = self.new_unconditional_stage(layer);
-        a_false.transfer = Some(Transfer::unconditional(false_stage.interface));
-        b_false.transfer = Some(Transfer::unconditional(false_stage.interface));
-        self.finish_stage(a_true);
-        self.finish_stage(a_false);
-        self.finish_stage(b_false);
-        swap_if(invert_out, (true_stage, false_stage))
-      }
-    }
-  }
-
-  fn _distill_fn<B>(
-    &mut self,
-    stage: &mut Stage,
-    fn_local: Local,
-    params: &[(Pat<'core>, Option<Ty<'core>>)],
-    body: &B,
-    distill_body: impl Fn(&mut Self, &mut Stage, &B) -> Port,
-  ) {
-    let return_local = self.new_local(stage);
-
-    let params = params.iter().map(|(p, _)| self.distill_pat_value(stage, p)).collect::<Vec<_>>();
-
-    let result = stage.take_local(return_local);
-    let wire = stage.new_wire();
-    stage.steps.push(Step::Fn(wire.0, params, result));
-    stage.set_local_to(fn_local, wire.1);
-
-    self.returns.push(Return { layer: stage.layer, local: return_local });
-
-    let result = distill_body(self, stage, body);
-    stage.set_local_to(return_local, result);
-
-    self.returns.pop();
   }
 
   fn distill_vec<T>(
@@ -808,16 +536,4 @@ impl<'core, 'r> Distiller<'core, 'r> {
 fn adt(path: &GenericPath) -> impl Fn(Port, Vec<Port>) -> Step {
   let adt = path.path.resolved.unwrap();
   move |x, y| Step::Adt(adt, x, y)
-}
-
-fn swap<T>((a, b): (T, T)) -> (T, T) {
-  (b, a)
-}
-
-fn swap_if<T>(bool: bool, (a, b): (T, T)) -> (T, T) {
-  if bool {
-    (b, a)
-  } else {
-    (a, b)
-  }
 }
