+++ vine/src/distiller/control_flow.rs
@@ -1,0 +1,349 @@
+use ivm::ext::ExtFnKind;
+
+use crate::{
+  ast::{Block, ComparisonOp, Expr, ExprKind, LabelId, Local, LogicalOp, Pat, PatKind, Span, Ty},
+  vir::{Interface, InterfaceKind, Layer, Port, Stage, Step, Transfer},
+};
+
+use super::{pattern_matching::Row, Distiller, Label, Return};
+
+impl<'core, 'r> Distiller<'core, 'r> {
+  pub(super) fn distill_match(
+    &mut self,
+    stage: &mut Stage,
+    value: &Expr<'core>,
+    arms: &Vec<(Pat<'core>, Expr<'core>)>,
+  ) -> Port {
+    let local = self.new_local(stage);
+    let (mut layer, mut init_stage) = self.child_layer(stage);
+    let value = self.distill_expr_value(&mut init_stage, value);
+    let rows = arms
+      .iter()
+      .map(|(pat, expr)| {
+        let mut stage = self.new_unconditional_stage(&mut layer);
+        let result = self.distill_expr_value(&mut stage, expr);
+        stage.set_local_to(local, result);
+        let interface = stage.interface;
+        self.finish_stage(stage);
+        Row::new(pat, interface)
+      })
+      .collect();
+    self.distill_pattern_match(&mut layer, &mut init_stage, value, rows);
+    self.finish_stage(init_stage);
+    self.finish_layer(layer);
+    stage.take_local(local)
+  }
+
+  pub(super) fn distill_fn(
+    &mut self,
+    stage: &mut Stage,
+    params: &Vec<(Pat<'core>, Option<Ty<'core>>)>,
+    body: &Expr<'core>,
+  ) -> Port {
+    let fn_local = self.new_local(stage);
+    let (layer, mut body_stage) = self.child_layer(stage);
+
+    self._distill_fn(&mut body_stage, fn_local, params, body, Self::distill_expr_value);
+
+    self.finish_stage(body_stage);
+    self.finish_layer(layer);
+
+    stage.take_local(fn_local)
+  }
+
+  pub(super) fn _distill_fn<B>(
+    &mut self,
+    stage: &mut Stage,
+    fn_local: Local,
+    params: &[(Pat<'core>, Option<Ty<'core>>)],
+    body: &B,
+    distill_body: impl Fn(&mut Self, &mut Stage, &B) -> Port,
+  ) {
+    let return_local = self.new_local(stage);
+
+    let params = params.iter().map(|(p, _)| self.distill_pat_value(stage, p)).collect::<Vec<_>>();
+
+    let result = stage.take_local(return_local);
+    let wire = stage.new_wire();
+    stage.steps.push(Step::Fn(wire.0, params, result));
+    stage.set_local_to(fn_local, wire.1);
+
+    self.returns.push(Return { layer: stage.layer, local: return_local });
+
+    let result = distill_body(self, stage, body);
+    stage.set_local_to(return_local, result);
+
+    self.returns.pop();
+  }
+
+  pub(super) fn distill_continue(&mut self, stage: &mut Stage, label: LabelId) -> Port {
+    let label = self.labels[label].as_ref().unwrap();
+    let transfer = Transfer::unconditional(label.continue_transfer.unwrap());
+    stage.steps.push(Step::Diverge(label.layer, Some(transfer)));
+    Port::Erase
+  }
+
+  pub(super) fn distill_break(
+    &mut self,
+    stage: &mut Stage,
+    label: LabelId,
+    value: &Option<Box<Expr<'core>>>,
+  ) -> Port {
+    let value = value.as_ref().map(|v| self.distill_expr_value(stage, v)).unwrap_or(Port::Erase);
+
+    let label = self.labels[label].as_ref().unwrap();
+    if let Some(local) = label.break_value {
+      stage.set_local_to(local, value);
+    } else {
+      stage.erase(value);
+    }
+
+    stage.steps.push(Step::Diverge(label.layer, None));
+
+    Port::Erase
+  }
+
+  pub(super) fn distill_return(
+    &mut self,
+    stage: &mut Stage,
+    value: &Option<Box<Expr<'core>>>,
+  ) -> Port {
+    let value = value.as_ref().map(|v| self.distill_expr_value(stage, v)).unwrap_or(Port::Erase);
+    let return_ = self.returns.last().unwrap();
+
+    stage.set_local_to(return_.local, value);
+    stage.steps.push(Step::Diverge(return_.layer, None));
+
+    Port::Erase
+  }
+
+  pub(super) fn distill_loop(
+    &mut self,
+    stage: &mut Stage,
+    label: LabelId,
+    block: &Block<'core>,
+  ) -> Port {
+    let local = self.new_local(stage);
+    let (layer, mut body_stage) = self.child_layer(stage);
+
+    *self.labels.get_or_extend(label) = Some(Label {
+      layer: layer.id,
+      continue_transfer: Some(body_stage.interface),
+      break_value: Some(local),
+    });
+
+    let result = self.distill_block(&mut body_stage, block);
+    body_stage.erase(result);
+    body_stage.transfer = Some(Transfer::unconditional(body_stage.interface));
+
+    self.finish_stage(body_stage);
+    self.finish_layer(layer);
+
+    stage.take_local(local)
+  }
+
+  pub(super) fn distill_while(
+    &mut self,
+    stage: &mut Stage,
+    label: LabelId,
+    cond: &Expr<'core>,
+    block: &Block<'core>,
+  ) -> Port {
+    let (mut layer, mut cond_stage) = self.child_layer(stage);
+
+    *self.labels.get_or_extend(label) = Some(Label {
+      layer: layer.id,
+      continue_transfer: Some(cond_stage.interface),
+      break_value: None,
+    });
+
+    let (mut then_stage, else_stage) = self.distill_cond(&mut layer, &mut cond_stage, cond);
+
+    let result = self.distill_block(&mut then_stage, block);
+    then_stage.erase(result);
+    then_stage.transfer = Some(Transfer::unconditional(cond_stage.interface));
+
+    self.finish_stage(cond_stage);
+    self.finish_stage(then_stage);
+    self.finish_stage(else_stage);
+    self.finish_layer(layer);
+
+    Port::Erase
+  }
+
+  pub(super) fn distill_if(
+    &mut self,
+    stage: &mut Stage,
+    cond: &Expr<'core>,
+    then_branch: &Block<'core>,
+    else_branch: &Expr<'core>,
+  ) -> Port {
+    let local = self.new_local(stage);
+    let (mut layer, mut cond_stage) = self.child_layer(stage);
+    let (mut then_stage, mut else_stage) = self.distill_cond(&mut layer, &mut cond_stage, cond);
+
+    let result = self.distill_block(&mut then_stage, then_branch);
+    then_stage.set_local_to(local, result);
+
+    let result = self.distill_expr_value(&mut else_stage, else_branch);
+    else_stage.set_local_to(local, result);
+
+    self.finish_stage(cond_stage);
+    self.finish_stage(then_stage);
+    self.finish_stage(else_stage);
+    self.finish_layer(layer);
+
+    stage.take_local(local)
+  }
+
+  pub(super) fn distill_do(
+    &mut self,
+    stage: &mut Stage,
+    label: LabelId,
+    block: &Block<'core>,
+  ) -> Port {
+    let local = self.new_local(stage);
+    let (layer, mut body_stage) = self.child_layer(stage);
+
+    *self.labels.get_or_extend(label) =
+      Some(Label { layer: layer.id, continue_transfer: None, break_value: Some(local) });
+
+    let result = self.distill_block(&mut body_stage, block);
+    body_stage.set_local_to(local, result);
+
+    self.finish_stage(body_stage);
+    self.finish_layer(layer);
+
+    stage.take_local(local)
+  }
+
+  pub(super) fn distill_cond_bool(
+    &mut self,
+    stage: &mut Stage,
+    cond: &Expr<'core>,
+    negate: bool,
+  ) -> Port {
+    if let ExprKind::Paren(inner) = &cond.kind {
+      return self.distill_cond_bool(stage, inner, negate);
+    }
+    match &cond.kind {
+      ExprKind![!cond] => {
+        let value = self.distill_expr_value(stage, cond);
+        if negate {
+          let wire = stage.new_wire();
+          stage.steps.push(Step::ExtFn(ExtFnKind::eq.into(), value, Port::N32(0), wire.0));
+          wire.1
+        } else {
+          value
+        }
+      }
+      ExprKind::Bool(b) => Port::N32((*b ^ negate) as u32),
+      ExprKind::Not(inner) => self.distill_cond_bool(stage, inner, !negate),
+      _ => {
+        let local = self.new_local(stage);
+        let (mut sub_layer, mut sub_stage) = self.child_layer(stage);
+        let (mut true_stage, mut false_stage) =
+          self.distill_cond(&mut sub_layer, &mut sub_stage, cond);
+        true_stage.set_local_to(local, Port::N32(if negate { 0 } else { 1 }));
+        false_stage.set_local_to(local, Port::N32(if negate { 1 } else { 0 }));
+        self.finish_stage(true_stage);
+        self.finish_stage(false_stage);
+        self.finish_stage(sub_stage);
+        self.finish_layer(sub_layer);
+        stage.take_local(local)
+      }
+    }
+  }
+
+  pub(super) fn distill_cond(
+    &mut self,
+    layer: &mut Layer,
+    stage: &mut Stage,
+    cond: &Expr<'core>,
+  ) -> (Stage, Stage) {
+    if let ExprKind::Paren(inner) = &cond.kind {
+      return self.distill_cond(layer, stage, inner);
+    }
+    if let ExprKind::ComparisonOp(lhs, cmps) = &cond.kind {
+      if let [(op, rhs)] = &**cmps {
+        if matches!(rhs.kind, ExprKind::N32(0)) {
+          if *op == ComparisonOp::Ne {
+            return self.distill_cond(layer, stage, lhs);
+          } else if *op == ComparisonOp::Eq {
+            return swap(self.distill_cond(layer, stage, lhs));
+          }
+        }
+      }
+    }
+    match &cond.kind {
+      ExprKind![!cond] => {
+        let bool = self.distill_expr_value(stage, cond);
+        let interface = self.interfaces.push(None);
+        let true_stage = self.new_stage(layer, interface);
+        let false_stage = self.new_stage(layer, interface);
+        self.interfaces[interface] = Some(Interface::new(
+          interface,
+          layer.id,
+          InterfaceKind::Branch([false_stage.id, true_stage.id]),
+        ));
+        stage.transfer = Some(Transfer { interface, data: Some(bool) });
+        (true_stage, false_stage)
+      }
+      ExprKind::Bool(bool) => {
+        let true_stage = self.new_unconditional_stage(layer);
+        let false_stage = self.new_unconditional_stage(layer);
+        stage.transfer = Some(Transfer::unconditional(if *bool {
+          true_stage.interface
+        } else {
+          false_stage.interface
+        }));
+        (true_stage, false_stage)
+      }
+      ExprKind::Not(inner) => swap(self.distill_cond(layer, stage, inner)),
+      ExprKind::Is(value, pat) => {
+        let value = self.distill_expr_value(stage, value);
+        let true_stage = self.new_unconditional_stage(layer);
+        let false_stage = self.new_unconditional_stage(layer);
+        self.distill_pattern_match(
+          layer,
+          stage,
+          value,
+          vec![
+            Row::new(pat, true_stage.interface),
+            Row::new(&Pat { span: Span::NONE, kind: PatKind::Hole }, false_stage.interface),
+          ],
+        );
+        (true_stage, false_stage)
+      }
+      ExprKind::LogicalOp(op, a, b) => {
+        let (invert_out, invert_a, invert_b) = match op {
+          LogicalOp::And => (false, false, false),
+          LogicalOp::Or => (true, true, true), // A || B == !(!A && !B)
+          LogicalOp::Implies => (true, false, true), // A => B == !(A && !B)
+        };
+        let (mut a_true, mut a_false) = swap_if(invert_a, self.distill_cond(layer, stage, a));
+        let (b_true, mut b_false) = swap_if(invert_b, self.distill_cond(layer, &mut a_true, b));
+        let true_stage = b_true;
+        let false_stage = self.new_unconditional_stage(layer);
+        a_false.transfer = Some(Transfer::unconditional(false_stage.interface));
+        b_false.transfer = Some(Transfer::unconditional(false_stage.interface));
+        self.finish_stage(a_true);
+        self.finish_stage(a_false);
+        self.finish_stage(b_false);
+        swap_if(invert_out, (true_stage, false_stage))
+      }
+    }
+  }
+}
+
+fn swap<T>((a, b): (T, T)) -> (T, T) {
+  (b, a)
+}
+
+fn swap_if<T>(bool: bool, (a, b): (T, T)) -> (T, T) {
+  if bool {
+    (b, a)
+  } else {
+    (a, b)
+  }
+}
